consolidate options controllers
